+++ knowledge_flow_backend/app/features/statistic/structures.py
@@ -12,86 +12,38 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import List, Literal, Optional, Any, Dict
-from enum import Enum
-from pydantic import BaseModel, model_validator
-
-# -- Constants for consistent types --
-DTypes = Literal["string", "integer", "float", "boolean", "datetime", "unknown"]
-
-# -- Schema models --
+from typing import List, Literal, Any, Dict
+from pydantic import BaseModel
 
 
-class TabularColumnSchema(BaseModel):
+class SetDatasetRequest(BaseModel):
+    dataset_name : str
+
+class DetectOutliersRequest(BaseModel):
+    method: Literal["zscore", "iqr"] = "zscore"
+    threshold: float = 3.0
+
+class CorrelationsRequest(BaseModel):
+    top_n: int = 5
+
+class PlotHistogramRequest(BaseModel):
+    column: str
+    bins: int = 30
+
+class PlotScatterRequest(BaseModel):
+    x_col: str
+    y_col: str
+
+class TrainModelRequest(BaseModel):
+    target: str
+    features: List[str]
+    model_type: Literal["linear", "random_forest"] = "linear"
+
+class PredictRowRequest(BaseModel):
+    row: Dict[str, Any]
+
+class SaveModelRequest(BaseModel):
     name: str
-    dtype: DTypes
 
-
-class TabularSchemaResponse(BaseModel):
-    document_name: str
-    columns: List[TabularColumnSchema]
-    row_count: Optional[int] = None
-
-
-class RawSQLRequest(BaseModel):
-    query: str
-
-
-class TabularQueryResponse(BaseModel):
-    db_name: str
-    sql_query: str
-    rows: Optional[List[dict]] = []
-    error: Optional[str] = None
-
-
-class HowToMakeAQueryResponse(BaseModel):
-    how: str
-
-
-class TabularDatasetMetadata(BaseModel):
-    document_name: str
-    title: str
-    description: Optional[str] = ""
-    tags: List[str] = []
-    domain: Optional[str] = ""
-    row_count: Optional[int] = None
-
-
-# -- Aggregation Models --
-
-
-class Precision(str, Enum):
-    sec = "sec"
-    min = "min"
-    hour = "hour"
-    day = "day"
-    all = "all"  # for global (non-bucketed) aggregations
-
-
-class Aggregation(str, Enum):
-    min = "MIN"
-    max = "MAX"
-    count = "COUNT"
-    count_distinct = "COUNT_DISTINCT"
-    avg = "AVG"
-    sum = "SUM"
-
-
-class AggregatedBucket(BaseModel):
-    time_bucket: str
-    values: Dict[str, Any]
-    groupby_fields: Optional[Dict[str, Any]] = None
-
-    @model_validator(mode="before")
-    @classmethod
-    def extract_groupby_fields(cls, values):
-        if "groupby_fields" not in values:
-            reserved = {"time_bucket", "values"}
-            groupby = {k: v for k, v in values.items() if k not in reserved}
-            if groupby:
-                values["groupby_fields"] = groupby
-        return values
-
-
-class TabularAggregationResponse(BaseModel):
-    buckets: List[AggregatedBucket]
+class LoadModelRequest(BaseModel):
+    name: str
