+++ knowledge_flow_backend/app/features/statistic/utils.py
@@ -1,5 +1,15 @@
-
+import math
 import datetime
+
+def clean_json(obj):
+    if isinstance(obj, dict):
+        return {k: clean_json(v) for k, v in obj.items()}
+    elif isinstance(obj, list):
+        return [clean_json(v) for v in obj]
+    elif isinstance(obj, float):
+        if math.isnan(obj) or math.isinf(obj):
+            return None  # ou str(obj) si tu veux garder "NaN", "inf"
+    return obj
 
 def safe_eval_function(func_str: str) -> callable:
     allowed_builtins = {
